pyproject.toml
[tool.poetry]
name = "term-paper-zaberov-drf-7"
version = "0.1.0"
description = ""
authors = ["zaberov.dv"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.8"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
ipython = "^8.23.0"
psycopg2-binary = "^2.9.9"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.4"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
stripe = "^8.10.0"
redis = "^5.0.3"
celery = "^5.3.6"
django-celery-beat = "^2.6.0"
eventlet = "^0.36.1"
django-extensions = "^3.2.3"
icecream = "^2.1.3"
pillow = "^10.3.0"
flake8-django = "^1.4"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
